Move search filters to sticky search navbar
