+++ frontend/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import Navbar from "@/components/navbar";
+import SearchNavbar from "@/components/search-navbar";
 import { DocumentsChartProvider } from "@/components/documents-chart-context";
 
 const geistSans = Geist({
@@ -29,7 +30,8 @@ export default function RootLayout({
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
         <DocumentsChartProvider>
           <Navbar />
-          <main className="pt-16">{children}</main>
+          <SearchNavbar />
+          <main>{children}</main>
         </DocumentsChartProvider>
       </body>
     </html>
